requirements.txt
python-telegram-bot==20.3
google-generativeai==0.4.0
textblob==0.17.1
pdf2image==1.16.3
pytesseract==0.3.10
pillow==10.0.1
aiohttp==3.8.1
pymongo==3.12.0
python-dotenv==0.19.2
fastapi==0.103.1
uvicorn==0.23.2
requests==2.31.0
async_lru==2.0.4